Restructure into abstract pb client to allow non-netty channels.
